+++ projects/storefrontlib/src/cms-components/cart/cart-shared/cart-item/cart-item.component.ts
@@ -16,10 +16,8 @@ import { Observable } from 'rxjs';
 import { PromotionService } from '../../../../shared/services/promotion/promotion.service';
 import { ICON_TYPE } from '../../../misc/icon/icon.model';
 import { CartItemComponentOutlets } from './model/cart-item-component.outlets';
-import {
-  CartItemContext,
-  CartItemContextSource,
-} from './model/cart-item.context';
+import { CartItemContextSource } from './model/cart-item-context-source.model';
+import { CartItemContext } from './model/cart-item-context.model';
 
 /**
  * @deprecated since 3.0 - use `OrderEntry` instead
