+++ projects/storefrontlib/src/cms-components/cart/cart-shared/cart-item/index.ts
@@ -1,3 +1,2 @@
 export * from './cart-item.component';
-export * from './model/cart-item-component.outlets';
-export * from './model/cart-item.context';
+export * from './model/index';
