+++ projects/storefrontlib/src/cms-components/cart/cart-shared/cart-item/model/cart-item-context-source.model.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 import { FormControl } from '@angular/forms';
 import { OrderEntry, PromotionLocation } from '@spartacus/core';
 import { CartItemComponentOptions } from '../cart-item.component';
-import { CartItemContextSource } from './cart-item.context';
+import { CartItemContextSource } from './cart-item-context-source.model';
 
 describe('CartItemContextSource', () => {
   let contextSource: CartItemContextSource;
