+++ projects/storefrontlib/src/cms-components/cart/cart-shared/cart-item/model/cart-item-context-source.model.ts
@@ -1,21 +1,9 @@
 import { Injectable } from '@angular/core';
 import { FormControl } from '@angular/forms';
 import { OrderEntry, PromotionLocation } from '@spartacus/core';
-import { Observable, ReplaySubject } from 'rxjs';
+import { ReplaySubject } from 'rxjs';
 import { CartItemComponentOptions } from '../cart-item.component';
-
-/**
- * Context for `CartItemComponent`.
- */
-@Injectable()
-export abstract class CartItemContext {
-  readonly compact$: Observable<boolean>;
-  readonly readonly$: Observable<boolean>;
-  readonly item$: Observable<OrderEntry>;
-  readonly quantityControl$: Observable<FormControl>;
-  readonly promotionLocation$: Observable<PromotionLocation>;
-  readonly options$: Observable<CartItemComponentOptions>;
-}
+import { CartItemContext } from './cart-item-context.model';
 
 /**
  * Context source for `CartItemComponent`.
