+++ projects/storefrontlib/src/cms-components/cart/cart-shared/cart-item/model/cart-item-context.model.ts
@@ -1,0 +1,18 @@
+import { Injectable } from '@angular/core';
+import { FormControl } from '@angular/forms';
+import { OrderEntry, PromotionLocation } from '@spartacus/core';
+import { Observable } from 'rxjs';
+import { CartItemComponentOptions } from '../cart-item.component';
+
+/**
+ * Context for `CartItemComponent`.
+ */
+@Injectable()
+export abstract class CartItemContext {
+  readonly compact$: Observable<boolean>;
+  readonly readonly$: Observable<boolean>;
+  readonly item$: Observable<OrderEntry>;
+  readonly quantityControl$: Observable<FormControl>;
+  readonly promotionLocation$: Observable<PromotionLocation>;
+  readonly options$: Observable<CartItemComponentOptions>;
+}
