+++ projects/storefrontlib/src/cms-components/cart/cart-shared/cart-item/model/index.ts
@@ -1,0 +1,3 @@
+export * from './cart-item-component-outlets.model';
+export * from './cart-item-context-source.model';
+export * from './cart-item-context.model';
